requirements.txt
click==8.1.7
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
imutils==0.5.4
numpy==1.26.3
opencv-python==4.9.0.80
Pillow==10.2.0
requests==2.31.0
paho.mqtt==1.6.1
